Update smb.conf
"In the default install, every file and directory created by guests will have root as owner and group. This will prevent the actual computer user from editing/deleting the files created by a samba user. We set the “guest account” mapping to “nobody” and force every file created via samba to belong to the sambashare group and be editable by it."

See https://www.fomfus.com/articles/how-to-public-samba-shares-in-manjaro